refactor paste from clipboard
