+++ src/components/InputWithScanner.tsx
@@ -1,11 +1,10 @@
 import { IonInput, IonText } from '@ionic/react'
 import InputContainer from './InputContainer'
-import PasteIcon from '../icons/Paste'
-import ScanIcon from '../icons/Scan'
-import XIcon from '../icons/X'
 import { useRef, useEffect } from 'react'
 import { pasteFromClipboard } from '../lib/clipboard'
 import { hapticLight } from '../lib/haptics'
+import Paste from './Paste'
+import { ClearButtonOnInput, ScanButtonOnInput } from './Button'
 
 interface InputWithScannerProps {
   error?: string
@@ -18,32 +17,6 @@ interface InputWithScannerProps {
   placeholder?: string
   validator?: (arg0: string) => boolean
   value?: string
-}
-
-const pillBase: React.CSSProperties = {
-  display: 'inline-flex',
-  alignItems: 'center',
-  gap: '0.3rem',
-  padding: '0.4rem 0.65rem',
-  borderRadius: '999px',
-  border: '1px solid var(--dark20)',
-  background: 'var(--dark05)',
-  cursor: 'pointer',
-  fontSize: '13px',
-  color: 'var(--dark80)',
-  whiteSpace: 'nowrap',
-  minHeight: '36px',
-  position: 'relative',
-  touchAction: 'manipulation',
-  userSelect: 'none',
-  WebkitTapHighlightColor: 'transparent',
-}
-
-// Expands tap target to 44px without increasing visual size
-const hitAreaStyle: React.CSSProperties = {
-  content: '""',
-  position: 'absolute',
-  inset: '-4px',
 }
 
 export default function InputWithScanner({
@@ -99,34 +72,11 @@ export default function InputWithScanner({
       >
         <IonText slot='end'>
           {hasValue ? (
-            <button
-              type='button'
-              onClick={handleClear}
-              aria-label='Clear address'
-              style={{
-                ...pillBase,
-                padding: '0.4rem 0.5rem',
-                minWidth: '44px',
-                background: 'none',
-                border: 'none',
-                color: 'var(--dark50)',
-              }}
-            >
-              <span style={hitAreaStyle} />
-              <XIcon />
-            </button>
+            <ClearButtonOnInput onClick={handleClear} />
           ) : (
             <div style={{ display: 'flex', gap: '0.35rem' }}>
-              <button type='button' onClick={handlePaste} aria-label='Paste address' style={pillBase}>
-                <span style={hitAreaStyle} />
-                <PasteIcon />
-                Paste
-              </button>
-              <button type='button' onClick={handleScan} aria-label='Scan QR code' style={pillBase}>
-                <span style={hitAreaStyle} />
-                <ScanIcon />
-                Scan QR
-              </button>
+              <Paste validator={validator} onPaste={handlePaste} />
+              <ScanButtonOnInput onClick={handleScan} />
             </div>
           )}
         </IonText>
